Add validation to all fields

+++ src/main/java/khorsun/springtest/MVC/controllers/PeopleController.java
@@ -1,11 +1,13 @@
 package khorsun.springtest.MVC.controllers;
 
 
+
 import khorsun.springtest.MVC.dao.PersonDAO;
 import khorsun.springtest.MVC.models.Person;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -13,6 +15,8 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+
+import javax.validation.Valid;
 
 
 @Controller
@@ -28,36 +32,41 @@ public class PeopleController {
     @GetMapping()
     public String index(Model model){
     model.addAttribute("people",personDAO.index());
-    return "/people/index";
+    return "people/index";
     }
  @GetMapping("/{id}")
     public String show(@PathVariable("id") int id, Model model){
     model.addAttribute("person",personDAO.show(id));
-    return "/people/show";
+    return "people/show";
     }
     @GetMapping("/new")
     public String create(@ModelAttribute("person")Person person){
-    return "/people/new";
+    return "people/new";
     }
     @PostMapping()
-    public String createNewPerson(@ModelAttribute("person") Person person){
+    public String createNewPerson(@ModelAttribute("person") @Valid Person person, BindingResult bindingResult){
+    if (bindingResult.hasErrors())
+        return "people/new";
     personDAO.create(person);
-    return "redirect:/people";
+    return "redirect:people";
     }
     @GetMapping("/{id}/edit")
     public String showEdit(@PathVariable("id") int id,Model model){
     model.addAttribute("person",personDAO.show(id));
-    return "/people/edit";
+    return "people/edit";
     }
     @PatchMapping("/{id}")
-    public String edit(@ModelAttribute("person")Person person,@PathVariable("id") int id){
+    public String edit(@ModelAttribute("person")@Valid Person person,
+                       BindingResult bindingResult,@PathVariable("id") int id){
+    if (bindingResult.hasErrors())
+        return "people/edit";
     personDAO.edit(person,id);
-    return "redirect:/people";
+    return "redirect:people";
     }
 
     @DeleteMapping("/{id}")
     public String delete(@PathVariable("id")int id){
     personDAO.delete(id);
-    return "redirect:/people";
+    return "redirect:people";
     }
 }
